business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/GetMemoryChannelData.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class GetMemoryChannelData implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+h ");
        messageString.append(command.getParameter("Channel"));
        messageString.append("\n");
        return messageString;
    }
}
